разбил js код на модули
